requirements.txt
# FastAPI и сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Telegram Bot
python-telegram-bot==20.7

# Конфигурация
python-dotenv==1.0.0

# Логирование и утилиты
pydantic==2.5.0
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Для будущих итераций (закомментировано)
# chromadb==0.4.18
